extract components into seperate files
