+++ src/types.ts
@@ -1,0 +1,6 @@
+export type Squares = "X"[] | "O"[] | null[];
+
+export type BoardState = {
+  squares: Squares;
+  xIsNext: boolean;
+};
